comments/requirements.txt
aspy.yaml==1.2.0
certifi==2019.3.9
cfgv==1.5.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
identify==1.4.1
idna==2.8
importlib-metadata==0.8
importlib-resources==1.0.2
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
nodeenv==1.3.3
openapi-codec==1.3.2
pre-commit==1.14.4
pytz==2018.9
PyYAML==5.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.3
zipp==0.3.3
